requirements.txt
twilio==8.10.0
schedule==1.2.0
python-dotenv==1.0.0 